Fix multiple process support

+++ datastore/migrations/20220323183151_sessions.js
@@ -8,7 +8,7 @@ exports.up = function(knex) {
     return knex.schema.createTable(table,t=>{
         t.increments('id').primary()
         t.string('name')
-        t.integer('accessedAt')
+        t.bigint('accessedAt')
     })
 };
